Final state
window state checks are commented out because it decided to misbehave.  The thermocouple method for getExtTemp is commented out as a second IOT board was used as the external temp.  The floats are cast to ints in the sprintf as I never re-figured out how to enable floats in the trustudio project setting